Make the regal rat littering less annoying
Simply put, some headass decided it was a good idea to make the rat produce FLAMMABLE, SLIPPERY OIL by default when it goes to spawn a mess. i've just decided, since it looks like blood anyways ingame, to just cut it down to the blood mess. if this breaks shit, i will be contemplating whether or not i should KILL myself.